refactor(omlx-mac): drop Sidebar struct + group helpers after TabView switch
`AppView` now mounts `TabView(.sidebarAdaptable)` (prior commit), so
none of the custom sidebar machinery is reachable from production code
anymore. Pruning the dead surface and renaming the file to reflect the
one symbol it still hosts.

Deleted (zero callers after the TabView switch):
- `struct Sidebar` — the View that previously mounted the section list
- `struct SidebarItem` — per-row renderer
- `struct SidebarGroupLabel` — group heading row
- `enum SidebarGroup` — the (Server / Models / Benchmark / General)
  grouping enum. Groups now live inline as TabSection blocks in AppView.
- `AppSection.gradient: [Color]` — only `SidebarItem` consumed it.
- `AppSection.group: SidebarGroup` — only used by the old group label
  resolution.

Kept:
- `AppSection` enum (renamed file Sidebar.swift → AppSection.swift) with
  `id`, `title`, `symbol`, `fillsContentArea`. All still referenced from
  AppView's Tabs and ContentScaffold.
- `Squircle` / `SquircleGradient` — still used by AboutScreen,
  ModelsScreen, ModelSettingsScreen for hero badges.

API compatibility audit: every API touched here (TabView, Tab(_:systemImage:value:content:),
TabSection, .tabViewStyle(.sidebarAdaptable)) is available since
macOS 15.0 — matches the project's deployment target. The macOS 26-only
glass APIs (glassEffect, Glass.tint, glassEffectTransition) remain
behind their existing `if #available(macOS 26.0, *)` gates in
Theme/Glass.swift.

Build SUCCEEDED; all 15 test suites pass.
